--- a/src/include/miopen/solver.hpp
+++ b/src/include/miopen/solver.hpp
@@ -741,10 +741,7 @@ struct PerformanceImplicitGemmBwdDataV4R1Xdlops
     PerformanceImplicitGemmBwdDataV4R1Xdlops(int, int, int, int, int, int, bool, bool);
 
     PerformanceImplicitGemmBwdDataV4R1Xdlops();
-    PerformanceImplicitGemmBwdDataV4R1Xdlops(bool)
-        : PerformanceImplicitGemmBwdDataV4R1Xdlops()
-    {
-    }
+    PerformanceImplicitGemmBwdDataV4R1Xdlops(bool) : PerformanceImplicitGemmBwdDataV4R1Xdlops() {}
 
     bool operator==(const PerformanceImplicitGemmBwdDataV4R1Xdlops& other) const;