--- a/src/kernels/composable_kernel/src/kernel_wrapper/gridwise_convolution_backward_data_implicit_gemm_v4r1_xdlops_gnchw_gkcyx_gnkhw.cpp
+++ b/src/kernels/composable_kernel/src/kernel_wrapper/gridwise_convolution_backward_data_implicit_gemm_v4r1_xdlops_gnchw_gkcyx_gnkhw.cpp
@@ -93,102 +93,58 @@ extern "C" __global__
     constexpr index_t GemmBBlockCopySrcDataPerRead_GemmN =
         CK_PARAM_TUNABLE_GEMM_B_BLOCK_COPY_SRC_DATA_PER_READ_GEMM_N;
 
-#if MIOPEN_USE_FP32
-    // A matrix
-    using GemmABlockCopyThreadSliceLengths_GemmG_GemmK_GemmM =
-        Sequence<1, GemmABlockCopyThreadSliceLengths_GemmK, GemmABlockCopyThreadSliceLengths_GemmM>;
-
-    using GemmABlockCopyThreadClusterLengths_GemmG_GemmK_GemmM =
-        Sequence<1, GemmABlockCopyClusterLengths_GemmK, GemmABlockCopyClusterLengths_GemmM>;
-
-    constexpr index_t GemmABlockCopyDstDataPerWrite_GemmM =
-        CK_PARAM_TUNABLE_GEMM_A_BLOCK_COPY_DST_DATA_PER_WRITE_GEMM_M;
-
-    // B matrix
-    using GemmBBlockCopyThreadSliceLengths_GemmG_GemmK_GemmN =
-        Sequence<1, GemmBBlockCopyThreadSliceLengths_GemmK, GemmBBlockCopyThreadSliceLengths_GemmN>;
-
-    using GemmBBlockCopyThreadClusterLengths_GemmG_GemmK_GemmN =
-        Sequence<1, GemmBBlockCopyClusterLengths_GemmK, GemmBBlockCopyClusterLengths_GemmN>;
-
-    constexpr index_t GemmBBlockCopyDstDataPerWrite_GemmN =
-        CK_PARAM_TUNABLE_GEMM_B_BLOCK_COPY_DST_DATA_PER_WRITE_GEMM_N;
-
-#elif MIOPEN_USE_FP16 || MIOPEN_USE_BFP16
     constexpr index_t GemmKPACK = CK_PARAM_KPACK_LENGTH;
 
+    constexpr index_t GemmABlockCopyClusterLengths_GemmKPACK =
+        CK_PARAM_DEPENDENT_GEMM_A_BLOCK_COPY_CLUSTER_LENGTHS_GEMM_KPACK;
+
+    constexpr index_t GemmBBlockCopyClusterLengths_GemmKPACK =
+        CK_PARAM_DEPENDENT_GEMM_B_BLOCK_COPY_CLUSTER_LENGTHS_GEMM_KPACK;
+
     // A matrix
+
+    constexpr index_t GemmABlockCopyThreadSliceLengths_GemmKPACK =
+        GemmKPACK / GemmABlockCopyClusterLengths_GemmKPACK;
+
     using GemmABlockCopyThreadSliceLengths_GemmG_GemmK_GemmM_GemmKPACK =
         Sequence<1,
                  GemmABlockCopyThreadSliceLengths_GemmK,
                  GemmABlockCopyThreadSliceLengths_GemmM,
-                 GemmKPACK>;
+                 GemmABlockCopyThreadSliceLengths_GemmKPACK>;
 
     using GemmABlockCopyThreadClusterLengths_GemmG_GemmK_GemmM_GemmKPACK =
-        Sequence<1, GemmABlockCopyClusterLengths_GemmK, GemmABlockCopyClusterLengths_GemmM, 1>;
+        Sequence<1,
+                 GemmABlockCopyClusterLengths_GemmK,
+                 GemmABlockCopyClusterLengths_GemmM,
+                 GemmABlockCopyClusterLengths_GemmKPACK>;
 
     constexpr index_t GemmABlockCopyDstDataPerWrite_GemmKPACK =
         CK_PARAM_TUNABLE_GEMM_A_BLOCK_COPY_DST_DATA_PER_WRITE_GEMM_KPACK;
 
-    // A matrix
+    // B matrix
+
+    constexpr index_t GemmBBlockCopyThreadSliceLengths_GemmKPACK =
+        GemmKPACK / GemmBBlockCopyClusterLengths_GemmKPACK;
+
     using GemmBBlockCopyThreadSliceLengths_GemmG_GemmK_GemmN_GemmKPACK =
         Sequence<1,
                  GemmBBlockCopyThreadSliceLengths_GemmK,
                  GemmBBlockCopyThreadSliceLengths_GemmN,
-                 GemmKPACK>;
+                 GemmBBlockCopyThreadSliceLengths_GemmKPACK>;
 
     using GemmBBlockCopyThreadClusterLengths_GemmG_GemmK_GemmN_GemmKPACK =
-        Sequence<1, GemmBBlockCopyClusterLengths_GemmK, GemmBBlockCopyClusterLengths_GemmN, 1>;
+        Sequence<1,
+                 GemmBBlockCopyClusterLengths_GemmK,
+                 GemmBBlockCopyClusterLengths_GemmN,
+                 GemmBBlockCopyClusterLengths_GemmKPACK>;
 
     constexpr index_t GemmBBlockCopyDstDataPerWrite_GemmKPACK =
         CK_PARAM_TUNABLE_GEMM_B_BLOCK_COPY_DST_DATA_PER_WRITE_GEMM_KPACK;
-
-#else
-    static_assert(false, "wrong! Only fp32, fp16 and bfp16 are supported.");
-#endif
 
     // C matrix
     constexpr auto GemmMPerWave = CK_PARAM_GEMM_M_PER_WAVE;
     constexpr auto GemmNPerWave = CK_PARAM_GEMM_N_PER_WAVE;
 
-    constexpr index_t GemmThreadGemmDataPerReadM = 1;
-    constexpr index_t GemmThreadGemmDataPerReadN = 1;
-
-#if MIOPEN_USE_FP32
-    constexpr auto gridwise_conv_bwd_data =
-        GridwiseConvolutionBackwardDataImplicitGemm_v4r1_xdlops_gnchw_gkcyx_gnkhw<
-            GridSize,
-            BlockSize,
-            FLOAT,
-            FLOAT_ACCUM,
-            decltype(in_gnchw_desc),
-            decltype(wei_gkcyx_desc),
-            decltype(out_gnkhw_desc),
-            ConvStrides,
-            ConvDilations,
-            InLeftPads,
-            InRightPads,
-            GemmMPerBlock,
-            GemmNPerBlock,
-            GemmKPerBlock,
-            GemmMPerWave,
-            GemmNPerWave,
-            GemmThreadGemmDataPerReadM,
-            GemmThreadGemmDataPerReadN,
-            GemmABlockCopyThreadSliceLengths_GemmG_GemmK_GemmM,
-            GemmABlockCopyThreadClusterLengths_GemmG_GemmK_GemmM,
-            GemmABlockCopySrcDataPerRead_GemmM,
-            GemmABlockCopyDstDataPerWrite_GemmM,
-            GemmBBlockCopyThreadSliceLengths_GemmG_GemmK_GemmN,
-            GemmBBlockCopyThreadClusterLengths_GemmG_GemmK_GemmN,
-            GemmBBlockCopySrcDataPerRead_GemmN,
-            GemmBBlockCopyDstDataPerWrite_GemmN>{};
-
-    // these decide which GEMM will be called
-    constexpr index_t GemmId = CK_PARAM_GEMM_ID;
-
-    gridwise_conv_bwd_data.template Run<GemmId>(p_in_global, p_wei_global, p_out_global);
-#elif MIOPEN_USE_FP16 || MIOPEN_USE_BFP16
     constexpr auto gridwise_conv_bwd_data =
         GridwiseConvolutionBackwardDataImplicitGemm_v4r1_xdlops_fp16_bfp16_gnchw_gkcyx_gnkhw<
             GridSize,
@@ -208,8 +164,6 @@ extern "C" __global__
             GemmKPACK,
             GemmMPerWave,
             GemmNPerWave,
-            GemmThreadGemmDataPerReadM,
-            GemmThreadGemmDataPerReadN,
             GemmABlockCopyThreadSliceLengths_GemmG_GemmK_GemmM_GemmKPACK,
             GemmABlockCopyThreadClusterLengths_GemmG_GemmK_GemmM_GemmKPACK,
             GemmABlockCopySrcDataPerRead_GemmM,
@@ -223,8 +177,4 @@ extern "C" __global__
     constexpr index_t GemmId = CK_PARAM_GEMM_ID;
 
     gridwise_conv_bwd_data.template Run<GemmId>(p_in_global, p_wei_global, p_out_global);
-
-#else
-    static_assert(false, "wrong! Only fp32, fp16 and bfp16 are supported.");
-#endif
 }