this kernel can support 256x128 block size and improve 3x3 performance