@asleepzzz 2020-09-07
[V4R1_perf.xlsx](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5182843/V4R1_perf.xlsx)