@atamazov 2020-09-07
@asleepzzz > [V4R1_perf.xlsx](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5182843/V4R1_perf.xlsx)

- (1) Could you please attach Find logs (taken with MIOPEN_LOG_LEVEL=5 and `-F 2`) for the cases where this new kernel wins?
- (2) The .xls file shows only 7 configs. Have you tested more?
- (3) The title of PR says that kernel is ___V4R1___. Excel table shows ConvHipImplicitGemmBwdData<b><i>V1R1</i></b>Xdlops. Why?