@atamazov 2020-09-07
@asleepzzz 
- (4) Does this PR require retuning of certain PerformanceConfgs?
- (5) How this kernel was tested for correctness?
- (6) Any regressions, i.e. cases when old kernel is faster than this one?
- (7) Does this change affect tuning time of this Solver?