@asleepzzz 2020-09-08
> @asleepzzz > [V4R1_perf.xlsx](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5182843/V4R1_perf.xlsx)
> 
> * (1) Could you please attach Find logs (taken with MIOPEN_LOG_LEVEL=5 and `-F 2`) for the cases when this kernel wins, prior and after this change (I see 3 such cases in .xlsx)?
> * (2) The .xlsx file shows only 7 configs. Have you tested more?
> * (3) The title of PR says that kernel is _**V4R1**_. Excel table shows ConvHipImplicitGemmBwdDataV1R1Xdlops. Why?

(2)this kernel is for non 1x1 configs,I also test 1x1 configs, not worse than V1R1 
(3)before this PR, MIOPEN choose V1R1