@asleepzzz 2020-09-09
[logs_old_v4r1.txt](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5194223/logs_old_v4r1.txt)
[logs_v4r1.txt](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5194225/logs_v4r1.txt)