@asleepzzz 2020-09-09
for ./bin/MIOpenDriver convfp16 -F 2 -n 128 -c 128 -H 58 -W 58 -k 128 -y 3 -x 3 -p 0 -q 0 -u 2 -v 2 -l 1 -j 1 -g 1 -t 1[
old version tuning time: 371sec
new version tuning time: 1114 sec
configs become more,but still much less than other solver