@asleepzzz 2020-09-10
[V4R1_perf (1).xlsx](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5199186/V4R1_perf.1.xlsx)

Even 1x1 stide1,V4R1 is better than V1R1 ,maybe we can change order of solver list......