@atamazov 2020-09-18
@asleepzzz 

>> * (1) Could you please attach Find logs (taken with MIOPEN_LOG_LEVEL=5 and `-F 2`) for the cases when this kernel wins, prior and after this change (I see 3 such cases in .xlsx)?

Please answer.

>> * (4) Does this PR require retuning of certain PerformanceConfgs?
>
> (4)yes, this kernel is not dynamic

"Dynamic" and "Tunable" are orthogonal notions.

I know that original Solver needs tuning. The question is about ** _ retuning _ **. Your PR changes existing Solver, so optimal values ​​of its PerformanceConfig may change. Or, format of records in performance database may change. If this is so, then we have to clearly inform Tuna guys.

If retuning is required, please notice that in the PR description and set appropriate label. Thanks.

>> * (5) How this kernel was tested for correctness?
>
> (5)pass the MIOpen verification

What is "MIOpen verification"? How can I reproduce these tests?

>> * (6) Any regressions, i.e. cases when old kernel is faster than this one?

Please answer.

>> * (7) Does this change affect tuning time of this Solver?

Please answer.