@asleepzzz 2020-09-23
[logs_old_v4r1.txt](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5265306/logs_old_v4r1.txt)
[logs_new_v4r1.txt](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5265307/logs_new_v4r1.txt)
Hi atamazov , these logs are old and new solver
average of tuning paremater: old is 9 ,new is 62
tuning time will increase due to we add more block size, we can use 256x128 now ,so we can better performance

for some cases,I don't know this called regression or not,most of them drop under 0.1%......
 
MIOpenDriver convfp16 -n 256 -c 512 -H 28 -W 28 -k 1024 -y 1 -x 1 -p 0 -q 0 -u 2 -v 2 -l 1 -j 1 -m conv -g 1 -F 2 -t 1
old:1.205ms
new:1.206ms 
drop 0.08%

MIOpenDriver convfp16 -n 256 -c 512 -H 7 -W 7 -k 512 -y 3 -x 3 -p 1 -q 1 -u 1 -v 1 -l 1 -j 1 -m conv -g 1 -F 2 -t 1
old:1.341ms
new:1.342ms
drop 0.07%

MIOpenDriver convfp16 -n 256 -c 64 -H 56 -W 56 -k 64 -y 1 -x 1 -p 0 -q 0 -u 1 -v 1 -l 1 -j 1 -m conv -g 1 -F 2 -t 1
old:0.2187ms
new:0.2190ms
drop 0.1%

MIOpenDriver convfp16 -n 256 -c 64 -H 56 -W 56 -k 64 -y 3 -x 3 -p 1 -q 1 -u 1 -v 1 -l 1 -j 1 -m conv -g 1 -F 2 -t 1
old:1.1989ms
new:1.199ms
drop 0.008%