@atamazov 2020-09-24
@asleepzzz Thanks for the logs. I am preparing some analysis now...