@atamazov 2020-09-24
> for some cases,I don't know this called regression or not,most of them drop under 0.1%...

This doesn't matter. Most likely this is due to jitter. Usually I ignore diffs < 1%. Diffs >5% worth close consideration.