@atamazov 2020-09-25
## Some questions about logs

> logs_old_v4r1.txt
> logs_new_v4r1.txt

Were these logs obtained with `MIOPEN_DEBUG_IMPLICIT_GEMM_FIND_ALL_SOLUTIONS=1`?

In the logs I see the following:
```
MIOpen(HIP): Warning [Monitor] 1241/0/1437 3.04385, best within recent 21: 4.1085 #1238 1,16,2,32,3,2,2,4, ETA:32.2508 sec.=====  204200
```
Please notice `=====  <decimal number>` at the end. Should I ignore these numbers?

It seems like these logs were collected with some code changes that are not included into this PR, right? What are those changes? Only logging?

## Log analysis.

With `MIOPEN_DEBUG_IMPLICIT_GEMM_FIND_ALL_SOLUTIONS=0`, the PR has no effect. The reason is that V1R1-xdlops precedes V4R1-xdlops in `GetImplicitGemmSolvers`.

With `MIOPEN_DEBUG_IMPLICIT_GEMM_FIND_ALL_SOLUTIONS=1`, V4R1-xdlops wins in 13 configs out of 21 total. The average performance gain is about of 4%. The drawback is ~5 hours of additional tuning time (15 minutes per config on average). Average tuning time slowdown of new V4R1-xdlops vs. old one is ~9X.

Note that it seems ___highly unlikely___ that we will use `MIOPEN_DEBUG_IMPLICIT_GEMM_FIND_ALL_SOLUTIONS=1` in the near future. Please see https://github.com/AMDComputeLibraries/MLOpen/pull/2520 for explanation.

Summary table (`MIOPEN_DEBUG_IMPLICIT_GEMM_FIND_ALL_SOLUTIONS=1`): [v4r1-xdlops-new-summary-table.xlsx](https://github.com/ROCmSoftwarePlatform/MIOpen/files/5279941/v4r1-xdlops-new-summary-table.xlsx)

## Conclusions:

- The PR wouldn't give any performance gain right away. However, I do not see any harm from merging it.
- Moreover, if it fixes #461, then it is definitely worth merging.
- The cost of this PR is necessity of re-tuning.
- I highly recommend considering optimization of V4R1-xdlops tuning time in the future.