@atamazov 2020-09-25
@asleepzzz
> > * (5) How this kernel was tested for correctness?
>
> (5)pass the MIOpen verification

Obviously ;) Could you please clarify how many configs have you tested with this Solver, i.e. with MIOPEN_DEBUG_FIND_ONLY_SOLVER=60? (Better, attach a list of that configs)