@atamazov 2020-09-25
W.R.T. re-tuning. At first, it is Ok to simply remove all the V4R1-xdlops PerformanceConfigs from the system-perf-db. The tuning "score" is about of ~1.1 on average which means that `HeuristicInit()` gives reasonably good default PerformanceConfigs.

/cc @JehandadKhan @daniellowell