@atamazov 2020-09-29
@asleepzzz 
> one log is only use V4R4 with this PR

V4R1, I guess

> another is use other solvers and do not have this PR

Usually I call this "baseline"

> This PR is for3x3 stride2,I will try to reduce tuning time

Good!